Feature/enums jpa

+++ src/main/java/org/sanmarcux/samples/sakila/controller/ActorRestController.java
@@ -81,7 +81,7 @@ public class ActorRestController {
      */
     @GetMapping("/{actorId}")
     @ResponseBody
-    public ActorDTO getActor(@PathVariable Short actorId) {
+    public ActorDTO getActor(@PathVariable Integer actorId) {
         LOG.info("Invoking Rest Service getActor");
         return actorBusiness.get(actorId);
     }
@@ -94,7 +94,7 @@ public class ActorRestController {
      * @return a modified actor
      */
     @PatchMapping("/{actorId}")
-    public ActorDTO modifyActor(@PathVariable final Short actorId, @RequestBody final ActorDTO payload) {
+    public ActorDTO modifyActor(@PathVariable final Integer actorId, @RequestBody final ActorDTO payload) {
         LOG.info("Invoking Rest Service modifyActor");
         return actorBusiness.modify(actorId, payload);
     }
@@ -105,7 +105,7 @@ public class ActorRestController {
      * @param actorId actor resource id
      */
     @DeleteMapping("/{actorId}")
-    public ResponseEntity<?> deleteActor(@PathVariable final Short actorId) {
+    public ResponseEntity<?> deleteActor(@PathVariable final Integer actorId) {
         LOG.info("Invoking Rest Service deleteActor");
         actorBusiness.delete(actorId);
 
@@ -113,7 +113,7 @@ public class ActorRestController {
     }
 
     @GetMapping("/{actorId}/films")
-    public ResponseEntity<List<FilmDTO>> listActorFilms(@PathVariable final Short actorId) {
+    public ResponseEntity<List<FilmDTO>> listActorFilms(@PathVariable final Integer actorId) {
         LOG.info("Invoking Rest Service listActorFilms");
         List<FilmDTO> films = filmBusiness.findFilmsByActor(actorId);
         if (films.isEmpty()) {
@@ -123,14 +123,14 @@ public class ActorRestController {
     }
 
     @GetMapping("/{actorId}/films/{filmId}")
-    public FilmDTO getActorFilm(@PathVariable final Short actorId, @PathVariable final Short filmId) {
+    public FilmDTO getActorFilm(@PathVariable final Integer actorId, @PathVariable final Integer filmId) {
         LOG.info("Invoking Rest Service getActorFilm");
 
         return actorBusiness.getFilm(actorId, filmId);
     }
 
     @PutMapping("/{actorId}/films/{filmId}")
-    public ResponseEntity<?> modifyActorFilm(@PathVariable final Short actorId, @PathVariable final Short filmId, @RequestBody final FilmDTO payload) {
+    public ResponseEntity<?> modifyActorFilm(@PathVariable final Integer actorId, @PathVariable final Integer filmId, @RequestBody final FilmDTO payload) {
         LOG.info("Invoking Rest Service modifyActorFilm");
 
         actorBusiness.createFilmParticipation(actorId, filmId);
@@ -139,7 +139,7 @@ public class ActorRestController {
     }
 
     @DeleteMapping("/{actorId}/films/{filmId}")
-    public ResponseEntity<?> deleteActorFilm(@PathVariable final Short actorId, @PathVariable final Short filmId) {
+    public ResponseEntity<?> deleteActorFilm(@PathVariable final Integer actorId, @PathVariable final Integer filmId) {
         LOG.info("Invoking Rest Service deleteActorFilm");
         actorBusiness.deleteFilm(actorId, filmId);
