+++ src/main/java/org/sanmarcux/samples/sakila/dto/FilmDTO.java
@@ -6,7 +6,6 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import java.math.BigDecimal;
 import java.time.Year;
-import java.util.Date;
 
 /**
  * Created on 22/04/2018.
@@ -15,7 +14,7 @@ import java.util.Date;
  */
 public class FilmDTO {
 
-    private Short filmId;
+    private Integer filmId;
     private String title;
     private String description;
     private Year releaseYear;
@@ -33,11 +32,11 @@ public class FilmDTO {
     private Rating rating;
     private String specialFeatures;
 
-    public Short getFilmId() {
+    public Integer getFilmId() {
         return filmId;
     }
 
-    public void setFilmId(Short filmId) {
+    public void setFilmId(Integer filmId) {
         this.filmId = filmId;
     }
